Implemented multithreading for vertex generation

+++ VertexArray.cpp
@@ -18,7 +18,8 @@ void VertexArray::AddBuffer(const VertexBuffer& vb, const VertexBufferLayout& la
     vb.Bind();
 
     const auto& elements = layout.GetElements();
-    unsigned int offset = 0;
+    // unsigned int offset = 0;
+    uintptr_t offset = 0;
 
     for(unsigned int i = 0; i < elements.size(); ++i)
     {
